[UPSTREAM] nano_systick: make nano timers robust to timer expiring in the past
When anouncing ticks to nanokernel, it updates ticks in timers. The assertion
is done that the tick anouncing is never late, that is to say no timer has
expired (hard real time assertion). When using deepsleep in ARC, sometimes, the
elapsed ticks anounced is too late, resulting on a missing of a timer expiring.
Its ticks underflows, leading to a never-expiring timer (or in a very long
time).

Fixed it by updating all expired timers, and verifying that the expiration
never underflows.

Change-Id: Ic9d1bc5f2fdeb35cf95aaf94256b8d293fadd027
Tracked-On: https://jira.ndg.intel.com/browse/FIRE-4330
Signed-off-by: David Coutherut <davidx.coutherut@intel.com>
Signed-off-by: Simon Desfarges <simon.desfarges@intel.com>
Reviewed-on: https://android.intel.com/471606
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Chereau, Fabien <fabien.chereau@intel.com>