[REVERTME] arc: call a hook before going idle
Also exposes variables needed by the hooks internals for
re-synchronizing the ARC ticks after a deepsleep period.

Note: the nano_cpu_atomic_idle saves the r0 register before calling
the idle_hook, and restores this register right after the call.
But the branch to _idle_exit label depends on return of idle_hook,
which is overwritten by the preceeding restore.

Fixed this behavior by saving the result of idle_hook in another
register (r2) and branch depending on this register.

Change-Id: I223e17c5eae11200495a664247fe150376138e1c
Tracked-On: https://jira.ndg.intel.com/browse/FIRE-4194
Signed-off-by: Simon Desfarges <simon.desfarges@intel.com>
Reviewed-on: https://android.intel.com/467028
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: D'alton, Alexandre <alexandre.dalton@intel.com>